pull spectator out to a manager
