+++ src/main/java/io/github/gaming32/uhcserver/managers/UHCStateManager.java
@@ -1,30 +1,24 @@
 package io.github.gaming32.uhcserver.managers;
 
-import net.minecraft.level.chunk.Chunk;
-import net.minecraft.packet.play.ChatMessagePacket;
-import net.minecraft.server.player.ServerPlayer;
 import io.github.gaming32.uhcserver.Formatting;
 import io.github.gaming32.uhcserver.UHCServerMod;
 import io.github.gaming32.uhcserver.UHCStages;
 import io.github.gaming32.uhcserver.access.IServerPlayer;
+import net.minecraft.level.chunk.Chunk;
+import net.minecraft.packet.play.ChatMessagePacket;
+import net.minecraft.server.player.ServerPlayer;
 
-import java.util.HashSet;
-import java.util.Set;
 import java.util.concurrent.ThreadLocalRandom;
 
 public class UHCStateManager {
 
     private UHCState uhcState = UHCState.WAITING;
     private UHCStages uhcStage = UHCStages.GRACE_PERIOD;
-    private final Set<String> spectators = new HashSet<>();
     private int tickLeft = uhcStage.getTime();
     public void endUHC() {
         uhcState = UHCState.ENDED;
-        synchronized (spectators) {
-            spectators.clear();
-        }
 
-        UHCServerMod.sendCustomPacket("reset-spectators");
+        UHCServerMod.getSpectatorManager().clear();
 
         UHCServerMod.getServer().levels[0].difficulty = 0;
         UHCServerMod.getServer().levels[1].difficulty = 0;
@@ -97,28 +91,14 @@ public class UHCStateManager {
         return uhcState == UHCState.RUNNING;
     }
 
-    public void sendAllSpectators(ServerPlayer player) {
-        synchronized (spectators) {
-            for (String p : spectators) {
-                player.packetHandler.send(new ChatMessagePacket("canyonuhc:spectator " + p));
-            }
-        }
-    }
 
     public void dead(String name) {
         if (uhcRunning()) {
-            setSpectator(name);
+            UHCServerMod.getSpectatorManager().setSpectator(name);
         }
     }
 
-    public void setSpectator(String name) {
-        spectators.add(name);
-        UHCServerMod.sendCustomPacket("spectator", name);
-    }
 
-    public boolean isSpectator(String name) {
-        return spectators.contains(name);
-    }
 
     private enum UHCState {
         WAITING,
